refactor: create separate package for common code for browser/react native sdk
