+++ packages/analytics-browser/src/config.ts
@@ -15,7 +15,7 @@ import {
   getQueryParams,
   SessionManager,
   FetchTransport,
-} from '@amplitude/analytics-browser-utils';
+} from '@amplitude/analytics-client-common';
 
 import { LocalStorage } from './storage/local-storage';
 import { XHRTransport } from './transports/xhr';
