+++ packages/analytics-client-common/jest.config.js
