+++ packages/analytics-client-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@amplitude/analytics-browser-utils",
+  "name": "@amplitude/analytics-client-common",
   "version": "0.0.0",
   "description": "",
   "author": "Amplitude Inc",
