+++ packages/analytics-client-common/src/attribution/constants.ts
