+++ packages/analytics-client-common/src/cookie-name.ts
