+++ packages/analytics-client-common/src/session-manager.ts
