+++ packages/analytics-client-common/test/plugins/identity.test.ts
@@ -1,6 +1,6 @@
 import { IdentityEventSender } from '../../src/plugins/identity';
 import { Config } from '@amplitude/analytics-types';
-import { getAnalyticsConnector } from '../../src/utils/analytics-connector';
+import { getAnalyticsConnector } from '../../src/analytics-connector';
 
 describe('identity', () => {
   describe('execute', () => {
