+++ packages/analytics-client-common/test/session-manager.test.ts
