+++ packages/analytics-client-common/tsconfig.es5.json
