+++ packages/analytics-client-common/tsconfig.esm.json
