+++ packages/analytics-client-common/tsconfig.json
