--- packages/analytics-react-native/src/utils/analytics-connector.ts
@@ -1,5 +1,0 @@
-import { AnalyticsConnector } from '@amplitude/analytics-connector';
-
-export const getAnalyticsConnector = (): AnalyticsConnector => {
-  return AnalyticsConnector.getInstance('$default_instance');
-};
