+++ packages/analytics-react-native/test/config.test.ts
@@ -2,8 +2,8 @@ import * as Config from '../src/config';
 import * as LocalStorageModule from '../src/storage/local-storage';
 import * as core from '@amplitude/analytics-core';
 import { LogLevel, Storage, UserSession } from '@amplitude/analytics-types';
-import * as BrowserUtils from '@amplitude/analytics-browser-utils';
-import { SessionManager, getCookieName, FetchTransport } from '@amplitude/analytics-browser-utils';
+import * as BrowserUtils from '@amplitude/analytics-client-common';
+import { SessionManager, getCookieName, FetchTransport } from '@amplitude/analytics-client-common';
 import { isWeb } from '../src/utils/platform';
 
 describe('config', () => {
